Hm tracker test dev (#384)
* update hm tracker test data

* fixed hm tracker test

* fixed some compling warnings

* fixed some compling warnings of hm tracker

* clean spaces

* fixed codetyle of some common files

* push test

+++ modules/perception/obstacle/lidar/tracker/hm_tracker/track_object_distance.cc
@@ -245,7 +245,7 @@ float TrackObjectDistance::ComputeHistogramDistance(const ObjectTrackPtr& track,
   }
 
   float histogram_dist = 0.0;
-  for (int i = 0; i < old_object_shape_features.size(); ++i) {
+  for (size_t i = 0; i < old_object_shape_features.size(); ++i) {
     histogram_dist += std::fabs(old_object_shape_features[i] -
       new_object_shape_features[i]);
   }
